separate cumulative code logic from component as a hook
